+++ examples/utils/typescript-versions-host/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @fluid-example/typescript-versions-host
 
+## 2.110.0
+
+Dependency updates only.
+
+## 2.103.0
+
+Dependency updates only.
+
+## 2.102.0
+
+Dependency updates only.
+
 ## 2.101.0
 
 Dependency updates only.